+++ spring-cloud-tencent-commons/src/main/java/com/tencent/cloud/common/metadata/config/MetadataAutoConfiguration.java
@@ -19,11 +19,8 @@
 package com.tencent.cloud.common.metadata.config;
 
 import com.tencent.cloud.common.metadata.StaticMetadataManager;
-import com.tencent.cloud.common.metadata.filter.gateway.MetadataFirstScgFilter;
 import com.tencent.cloud.common.spi.InstanceMetadataProvider;
 
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.cloud.gateway.filter.GlobalFilter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.lang.Nullable;
@@ -51,16 +48,4 @@ public class MetadataAutoConfiguration {
 		return new StaticMetadataManager(metadataLocalProperties, instanceMetadataProvider);
 	}
 
-	/**
-	 * Create when gateway application is SCG.
-	 */
-	@Configuration
-	@ConditionalOnClass(name = "org.springframework.cloud.gateway.filter.GlobalFilter")
-	static class MetadataScgFilterConfig {
-
-		@Bean
-		public GlobalFilter metadataFirstScgFilter() {
-			return new MetadataFirstScgFilter();
-		}
-	}
 }