+++ CHANGELOG.md
@@ -3,4 +3,8 @@
 
 - [Bugfix: optimize ratelimit actuator](https://github.com/Tencent/spring-cloud-tencent/pull/413)
 - [Feature: add rate limit filter debug log](https://github.com/Tencent/spring-cloud-tencent/pull/417)
+- [Feature: Optimized configuration update](https://github.com/Tencent/spring-cloud-tencent/pull/423)
+- [Feature: add feature-env plugin & add spring cloud gateway staining plugin](https://github.com/Tencent/spring-cloud-tencent/pull/428)
+- [Optimize: add EncodeTransferMedataRestTemplateInterceptor to RestTemplate](https://github.com/Tencent/spring-cloud-tencent/pull/434)
+- [Optimize: Specification apollo code reference notes](https://github.com/Tencent/spring-cloud-tencent/pull/442)
 - [Feature: graceful service registration after ApplicationReadyEventg](https://github.com/Tencent/spring-cloud-tencent/pull/441)